move Datapoint out of public namespace
